@@ -19,58 +19,60 @@
     "generate": "ng generate"
   },
   "dependencies": {
-    "@angular/animations": "^10.1.3",
-    "@angular/cdk": "^10.2.2",
-    "@angular/common": "^10.1.3",
-    "@angular/compiler": "^10.1.3",
-    "@angular/core": "^10.1.3",
-    "@angular/forms": "^10.1.3",
-    "@angular/localize": "^10.1.3",
-    "@angular/platform-browser": "^10.1.3",
-    "@angular/platform-browser-dynamic": "^10.1.3",
-    "@angular/router": "^10.1.3",
+    "@angular/animations": "^10.2.4",
+    "@angular/cdk": "^10.2.7",
+    "@angular/common": "^10.2.4",
+    "@angular/compiler": "^10.2.4",
+    "@angular/core": "^10.2.4",
+    "@angular/forms": "^10.2.4",
+    "@angular/localize": "^10.2.4",
+    "@angular/platform-browser": "^10.2.4",
+    "@angular/platform-browser-dynamic": "^10.2.4",
+    "@angular/router": "^10.2.4",
     "@fortawesome/fontawesome-free": "^5.15.1",
     "@ng-bootstrap/ng-bootstrap": "^8.0.0",
     "@ng-select/ng-option-highlight": "0.0.5",
-    "@ng-select/ng-select": "^5.0.9",
+    "@ng-select/ng-select": "^5.0.14",
     "@ngx-translate/core": "^13.0.0 ",
     "@swimlane/ngx-charts": "^16.0.0",
     "@swimlane/ngx-datatable": "^18.0.0",
     "@types/file-saver": "^2.0.1",
     "angular-highcharts": "^10.0.1",
     "bootstrap": "^4.5.0",
-    "browserslist": "^4.8.6",
+    "browserslist": "^4.16.0",
     "classlist.js": "^1.1.20150312",
-    "core-js": "^3.6.4",
-    "file-saver": "^2.0.2",
-    "highcharts": "^8.1.2",
+    "core-js": "^3.8.1",
+    "file-saver": "^2.0.5",
+    "highcharts": "^8.2.2",
+    "mousetrap": "^1.6.5",
     "ngx-color-picker": "^10.1.0",
     "ngx-filesaver": "^11.0.0",
     "ngx-json-viewer": "^2.4.0",
+    "ngx-mousetrap": "^1.0.0",
     "ngx-perfect-scrollbar": "^10.0.1",
     "ngx-spinner": "^10.0.1",
-    "ngx-toastr": "^13.1.0",
+    "ngx-toastr": "^13.2.0",
     "rxjs": "^6.5.5",
-    "subsink": "^1.0.1",
+    "subsink": "^1.0.2",
     "toppy": "^2.3.4",
-    "tslib": "^2.0.0",
+    "tslib": "^2.0.3",
     "web-animations-js": "^2.3.2",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.2",
-    "@angular/cli": "^10.1.3",
-    "@angular/compiler-cli": "^10.1.3",
-    "@angular/language-service": "^10.1.3",
+    "@angular-devkit/build-angular": "^0.1100.5",
+    "@angular/cli": "^10.2.1",
+    "@angular/compiler-cli": "^10.2.4",
+    "@angular/language-service": "^10.2.4",
     "@biesbjerg/ngx-translate-extract": "^4.2.0",
     "@ngx-rocket/scripts": "^4.0.0",
-    "@types/jasmine": "^3.5.3",
+    "@types/jasmine": "^3.6.2",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^13.7.0",
+    "@types/node": "^13.13.38",
     "codelyzer": "^6.0.1",
     "htmlhint": "^0.11.0",
     "https-proxy-agent": "^5.0.0",
-    "husky": "^4.2.1",
+    "husky": "^4.3.6",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~5.0.0",
@@ -78,20 +80,20 @@
     "karma-cli": "~2.0.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-jasmine-html-reporter": "^1.5.4",
     "karma-junit-reporter": "^2.0.1",
     "prettier": "^1.19.1",
-    "pretty-quick": "^2.0.1",
+    "pretty-quick": "^2.0.2",
     "protractor": "~7.0.0",
     "stylelint": "~13.1.0",
-    "stylelint-config-prettier": "^8.0.1",
+    "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-recommended-scss": "~4.2.0",
     "stylelint-config-standard": "~20.0.0",
     "stylelint-scss": "~3.14.2",
-    "ts-node": "^8.6.2",
-    "tslint": "~6.1.0",
+    "ts-node": "^8.10.2",
+    "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "~4.0.3"
+    "typescript": "^4.0.5"
   },
   "prettier": {
     "singleQuote": true,
