@@ -4,12 +4,14 @@ import { ApiService } from '@app/services/api.service';
 import { VersionService } from '@app/services/version-service';
 import { TranslateService } from '@ngx-translate/core';
 import { Chart } from 'angular-highcharts';
-import { forkJoin } from 'rxjs';
+import { forkJoin, timer } from 'rxjs';
 import { GlobalPosition, InsidePlacement, Toppy, ToppyControl } from 'toppy';
 import { PluginStats } from '../shared/models/plugin-stats';
 import { DeviceByNameComponent } from './device-by-name/device-by-name.component';
 import { UnsubscribeOnDestroyAdapter } from '@app/shared/adapter/unsubscribe-adapter';
 import { HeaderService } from '@app/services/header-service';
+import { environment } from '@env/environment';
+import { switchMap, retry, share, takeUntil, filter, map } from 'rxjs/operators';
 
 const log = new Logger('DashboardComponent');
 
@@ -19,6 +21,7 @@ const log = new Logger('DashboardComponent');
   styleUrls: ['./dashboard.component.scss']
 })
 export class DashboardComponent extends UnsubscribeOnDestroyAdapter implements OnInit, OnDestroy {
+  poll = false;
   chart: Chart;
   devices: any;
   routers: any;
@@ -91,6 +94,23 @@ export class DashboardComponent extends UnsubscribeOnDestroyAdapter implements O
   }
 
   ngOnInit() {
+    this.subs.sink = this.headerService.polling.subscribe(poll => {
+      this.poll = poll;
+      this.subs.add(
+        timer(1, environment.refresh)
+          .pipe(
+            switchMap(() => this.getStats()),
+            retry(),
+            share(),
+            takeUntil(this.headerService.polling.pipe(filter(val => val === false)))
+          )
+          .subscribe()
+      );
+    });
+
+    if (!this.poll) {
+      this.getStats().subscribe();
+    }
     this.getInfos();
   }
 
@@ -102,137 +122,142 @@ export class DashboardComponent extends UnsubscribeOnDestroyAdapter implements O
     this.advancedPieDeviceLabel = this.translateService.instant('dashboard.devices.label');
     this.advancedPieStateLabel = this.translateService.instant('dashboard.devices.state.label');
     this.advancedPieBatteryLabel = this.translateService.instant('dashboard.devices.battery.label');
-    forkJoin([this.apiService.getPluginStats(), this.apiService.getZDevices()]).subscribe(([res, devices]) => {
-      this.pluginStats = res;
-      this.createChart();
-      this.totalTraficSent.label = this.translateService.instant('dashboard.trafic.total.trafic.sent');
-      this.totalTraficSent.total = res.Sent;
-      this.totalTraficRetx.label = this.translateService.instant('dashboard.trafic.retx');
-      this.totalTraficRetx.total = ((res.ReTx / res.Sent) * 100).toFixed(0);
-      this.totalTraficRetx.append = '%';
-      this.totalTraficReceived.label = this.translateService.instant('dashboard.trafic.total.trafic.received');
-      this.totalTraficReceived.total = res.Received;
-      this.totalTraficCluster.label = this.translateService.instant('dashboard.trafic.cluster');
-      this.totalTraficCluster.total = ((res.Cluster / res.Received) * 100).toFixed(0);
-      this.totalTraficCluster.append = '%';
-      this.totalTraficCrc.label = this.translateService.instant('dashboard.trafic.crc');
-      this.totalTraficCrc.total = ((res.CRC / res.Received) * 100).toFixed(0);
-      this.totalTraficCrc.append = '%';
-      this.maxLoad.label = this.translateService.instant('dashboard.trafic.maxload');
-      this.maxLoad.total = res.MaxLoad;
-      this.currentLoad.label = this.translateService.instant('dashboard.trafic.currentload');
-      this.currentLoad.total = res.CurrentLoad;
-      this.advancedPieLoad = [
-        { name: this.translateService.instant('dashboard.trafic.maxload'), value: res.MaxLoad - res.CurrentLoad },
-        { name: this.translateService.instant('dashboard.trafic.currentload'), value: res.CurrentLoad }
-      ];
-      this.advancedPieSent = [
-        { name: this.translateService.instant('dashboard.trafic.total.trafic.sent'), value: res.Sent - res.ReTx },
-        { name: this.translateService.instant('dashboard.trafic.retx'), value: res.ReTx }
-      ];
-      this.advancedPieReceived = [
-        {
-          name: this.translateService.instant('dashboard.trafic.total.trafic.received'),
-          value: res.Received - res.Cluster - res.CRC
-        },
-        { name: this.translateService.instant('dashboard.trafic.cluster'), value: res.Cluster },
-        { name: this.translateService.instant('dashboard.trafic.crc'), value: res.CRC }
-      ];
-      this.devices = devices;
-      this.devices.total = this.devices.length;
-      this.devices.label = this.translateService.instant('devices');
-      this.routers = this.devices.filter((router: any) => router.LogicalType === 'Router');
-      this.routers.label = this.translateService.instant('dashboard.devices.routers');
-      this.routers.total = ((this.routers.length / this.devices.total) * 100).toFixed(0);
-      this.routers.append = '%';
-      this.inDbs = this.devices.filter((router: any) => router.Status === 'inDB');
-      this.inDbs.label = this.translateService.instant('dashboard.devices.indb');
-      this.inDbs.total = ((this.inDbs.length / this.devices.total) * 100).toFixed(0);
-      this.inDbs.append = '%';
-      this.healthsLive = this.devices.filter((device: any) => device.Health === 'Live' && device.Status !== 'notDB');
-      this.healthsLive.label = this.translateService.instant('dashboard.devices.live');
-      this.healthsLive.total = ((this.healthsLive.length / this.devices.total) * 100).toFixed(0);
-      this.healthsLive.append = '%';
-      this.healthsNotReachable = this.devices.filter((device: any) => {
-        return device.Health === 'Not Reachable' && device.Status !== 'notDB';
-      });
-      this.healthsNotReachable.label = this.translateService.instant('dashboard.devices.notReachable');
-      this.healthsNotReachable.total = ((this.healthsNotReachable.length / this.devices.total) * 100).toFixed(0);
-      this.healthsNotReachable.append = '%';
-      this.healthsNotSeen = this.devices.filter((device: any) => {
-        return device.Health === 'Not seen last 24hours' && device.Status !== 'notDB';
-      });
-      this.healthsNotSeen.label = this.translateService.instant('dashboard.devices.notseen');
-      this.healthsNotSeen.total = ((this.healthsNotSeen.length / this.devices.total) * 100).toFixed(0);
-      this.healthsNotSeen.append = '%';
-      this.healthsOthers = this.devices.filter((device: any) => {
-        return (
-          device.Health !== 'Not seen last 24hours' &&
-          device.Health !== 'Not Reachable' &&
-          device.Health !== 'Live' &&
-          device.Status !== 'notDB'
+  }
+
+  private getStats() {
+    return forkJoin([this.apiService.getPluginStats(), this.apiService.getZDevices()]).pipe(
+      map(([res, devices]) => {
+        this.pluginStats = res;
+        this.createChart();
+        this.totalTraficSent.label = this.translateService.instant('dashboard.trafic.total.trafic.sent');
+        this.totalTraficSent.total = res.Sent;
+        this.totalTraficRetx.label = this.translateService.instant('dashboard.trafic.retx');
+        this.totalTraficRetx.total = ((res.ReTx / res.Sent) * 100).toFixed(0);
+        this.totalTraficRetx.append = '%';
+        this.totalTraficReceived.label = this.translateService.instant('dashboard.trafic.total.trafic.received');
+        this.totalTraficReceived.total = res.Received;
+        this.totalTraficCluster.label = this.translateService.instant('dashboard.trafic.cluster');
+        this.totalTraficCluster.total = ((res.Cluster / res.Received) * 100).toFixed(0);
+        this.totalTraficCluster.append = '%';
+        this.totalTraficCrc.label = this.translateService.instant('dashboard.trafic.crc');
+        this.totalTraficCrc.total = ((res.CRC / res.Received) * 100).toFixed(0);
+        this.totalTraficCrc.append = '%';
+        this.maxLoad.label = this.translateService.instant('dashboard.trafic.maxload');
+        this.maxLoad.total = res.MaxLoad;
+        this.currentLoad.label = this.translateService.instant('dashboard.trafic.currentload');
+        this.currentLoad.total = res.CurrentLoad;
+        this.advancedPieLoad = [
+          { name: this.translateService.instant('dashboard.trafic.maxload'), value: res.MaxLoad - res.CurrentLoad },
+          { name: this.translateService.instant('dashboard.trafic.currentload'), value: res.CurrentLoad }
+        ];
+        this.advancedPieSent = [
+          { name: this.translateService.instant('dashboard.trafic.total.trafic.sent'), value: res.Sent - res.ReTx },
+          { name: this.translateService.instant('dashboard.trafic.retx'), value: res.ReTx }
+        ];
+        this.advancedPieReceived = [
+          {
+            name: this.translateService.instant('dashboard.trafic.total.trafic.received'),
+            value: res.Received - res.Cluster - res.CRC
+          },
+          { name: this.translateService.instant('dashboard.trafic.cluster'), value: res.Cluster },
+          { name: this.translateService.instant('dashboard.trafic.crc'), value: res.CRC }
+        ];
+        this.devices = devices;
+        this.devices.total = this.devices.length;
+        this.devices.label = this.translateService.instant('devices');
+        this.routers = this.devices.filter((router: any) => router.LogicalType === 'Router');
+        this.routers.label = this.translateService.instant('dashboard.devices.routers');
+        this.routers.total = ((this.routers.length / this.devices.total) * 100).toFixed(0);
+        this.routers.append = '%';
+        this.inDbs = this.devices.filter((router: any) => router.Status === 'inDB');
+        this.inDbs.label = this.translateService.instant('dashboard.devices.indb');
+        this.inDbs.total = ((this.inDbs.length / this.devices.total) * 100).toFixed(0);
+        this.inDbs.append = '%';
+        this.healthsLive = this.devices.filter((device: any) => device.Health === 'Live' && device.Status !== 'notDB');
+        this.healthsLive.label = this.translateService.instant('dashboard.devices.live');
+        this.healthsLive.total = ((this.healthsLive.length / this.devices.total) * 100).toFixed(0);
+        this.healthsLive.append = '%';
+        this.healthsNotReachable = this.devices.filter((device: any) => {
+          return device.Health === 'Not Reachable' && device.Status !== 'notDB';
+        });
+        this.healthsNotReachable.label = this.translateService.instant('dashboard.devices.notReachable');
+        this.healthsNotReachable.total = ((this.healthsNotReachable.length / this.devices.total) * 100).toFixed(0);
+        this.healthsNotReachable.append = '%';
+        this.healthsNotSeen = this.devices.filter((device: any) => {
+          return device.Health === 'Not seen last 24hours' && device.Status !== 'notDB';
+        });
+        this.healthsNotSeen.label = this.translateService.instant('dashboard.devices.notseen');
+        this.healthsNotSeen.total = ((this.healthsNotSeen.length / this.devices.total) * 100).toFixed(0);
+        this.healthsNotSeen.append = '%';
+        this.healthsOthers = this.devices.filter((device: any) => {
+          return (
+            device.Health !== 'Not seen last 24hours' &&
+            device.Health !== 'Not Reachable' &&
+            device.Health !== 'Live' &&
+            device.Status !== 'notDB'
+          );
+        });
+        this.healthsOthers.label = this.translateService.instant('dashboard.devices.enddevice');
+        this.healthsOthers.total = ((this.healthsOthers.length / this.devices.total) * 100).toFixed(0);
+        this.healthsOthers.append = '%';
+        this.advancedPieDevice = [
+          {
+            name: this.translateService.instant('dashboard.devices.routers'),
+            value: this.routers.length
+          },
+          {
+            name: this.translateService.instant('dashboard.devices.enddevice'),
+            value: this.devices.length - this.routers.length
+          }
+        ];
+        this.advancedPieState = [
+          {
+            name: this.translateService.instant('dashboard.devices.live'),
+            value: this.healthsLive.length
+          },
+          {
+            name: this.translateService.instant('dashboard.devices.notReachable'),
+            value: this.healthsNotReachable.length
+          },
+          {
+            name: this.translateService.instant('dashboard.devices.notseen'),
+            value: this.healthsNotSeen.length
+          },
+          { name: this.translateService.instant('dashboard.devices.others'), value: this.healthsOthers.length }
+        ];
+        this.devicesOnBattery = devices.filter(
+          (device: any) => device.LogicalType !== 'Router' && device.Status !== 'notDB'
         );
-      });
-      this.healthsOthers.label = this.translateService.instant('dashboard.devices.enddevice');
-      this.healthsOthers.total = ((this.healthsOthers.length / this.devices.total) * 100).toFixed(0);
-      this.healthsOthers.append = '%';
-      this.advancedPieDevice = [
-        {
-          name: this.translateService.instant('dashboard.devices.routers'),
-          value: this.routers.length
-        },
-        {
-          name: this.translateService.instant('dashboard.devices.enddevice'),
-          value: this.devices.length - this.routers.length
-        }
-      ];
-      this.advancedPieState = [
-        {
-          name: this.translateService.instant('dashboard.devices.live'),
-          value: this.healthsLive.length
-        },
-        {
-          name: this.translateService.instant('dashboard.devices.notReachable'),
-          value: this.healthsNotReachable.length
-        },
-        {
-          name: this.translateService.instant('dashboard.devices.notseen'),
-          value: this.healthsNotSeen.length
-        },
-        { name: this.translateService.instant('dashboard.devices.others'), value: this.healthsOthers.length }
-      ];
-      this.devicesOnBattery = devices.filter(
-        (device: any) => device.LogicalType !== 'Router' && device.Status !== 'notDB'
-      );
-      const _batteryInf30 = this.devicesOnBattery.filter((device: any) => device.Battery < 30);
-      const _batterySup30 = this.devicesOnBattery.filter((device: any) => device.Battery > 30 && device.Battery < 50);
-      const _batterySup50 = this.devicesOnBattery.filter((device: any) => device.Battery > 50);
-      this.batteryInf30 = this.devices.filter((it: any) => _batteryInf30.find((iter: any) => iter.IEEE === it.IEEE));
-      this.batterySup50 = this.devices.filter((it: any) => _batterySup50.find((iter: any) => iter.IEEE === it.IEEE));
-      this.batterySup30 = this.devices.filter((it: any) => _batterySup30.find((iter: any) => iter.IEEE === it.IEEE));
-      this.batteryInf30.totalDevices = this.devicesOnBattery.length;
-      this.batteryInf30.label = this.translateService.instant('dashboard.devices.battery.inf.30');
-      this.batteryInf30.total = this.batteryInf30.length;
-      this.batteryInf30.append = this.translateService.instant('devices');
-      this.batterySup30.label = this.translateService.instant('dashboard.devices.battery.sup.30');
-      this.batterySup30.total = this.batterySup30.length;
-      this.batterySup30.append = this.translateService.instant('devices');
-      this.batterySup50.label = this.translateService.instant('dashboard.devices.battery.sup.50');
-      this.batterySup50.total = this.batterySup50.length;
-      this.batterySup50.append = this.translateService.instant('devices');
-      this.advancedPieBattery = [
-        {
-          name: this.translateService.instant('dashboard.devices.battery.inf.30'),
-          value: this.batteryInf30.length
-        },
-        {
-          name: this.translateService.instant('dashboard.devices.battery.sup.30'),
-          value: this.batterySup30.length
-        },
-        { name: this.translateService.instant('dashboard.devices.battery.sup.50'), value: this.batterySup50.length }
-      ];
-      this.headerService.setError(res.Error);
-    });
+        const _batteryInf30 = this.devicesOnBattery.filter((device: any) => device.Battery < 30);
+        const _batterySup30 = this.devicesOnBattery.filter((device: any) => device.Battery > 30 && device.Battery < 50);
+        const _batterySup50 = this.devicesOnBattery.filter((device: any) => device.Battery > 50);
+        this.batteryInf30 = this.devices.filter((it: any) => _batteryInf30.find((iter: any) => iter.IEEE === it.IEEE));
+        this.batterySup50 = this.devices.filter((it: any) => _batterySup50.find((iter: any) => iter.IEEE === it.IEEE));
+        this.batterySup30 = this.devices.filter((it: any) => _batterySup30.find((iter: any) => iter.IEEE === it.IEEE));
+        this.batteryInf30.totalDevices = this.devicesOnBattery.length;
+        this.batteryInf30.label = this.translateService.instant('dashboard.devices.battery.inf.30');
+        this.batteryInf30.total = this.batteryInf30.length;
+        this.batteryInf30.append = this.translateService.instant('devices');
+        this.batterySup30.label = this.translateService.instant('dashboard.devices.battery.sup.30');
+        this.batterySup30.total = this.batterySup30.length;
+        this.batterySup30.append = this.translateService.instant('devices');
+        this.batterySup50.label = this.translateService.instant('dashboard.devices.battery.sup.50');
+        this.batterySup50.total = this.batterySup50.length;
+        this.batterySup50.append = this.translateService.instant('devices');
+        this.advancedPieBattery = [
+          {
+            name: this.translateService.instant('dashboard.devices.battery.inf.30'),
+            value: this.batteryInf30.length
+          },
+          {
+            name: this.translateService.instant('dashboard.devices.battery.sup.30'),
+            value: this.batterySup30.length
+          },
+          { name: this.translateService.instant('dashboard.devices.battery.sup.50'), value: this.batterySup50.length }
+        ];
+        this.headerService.setError(res.Error);
+      })
+    );
   }
 
   percentageFormatting(value: any) {
@@ -404,7 +429,7 @@ export class DashboardComponent extends UnsubscribeOnDestroyAdapter implements O
   }
 
   refresh() {
-    this.getInfos();
+    this.getStats().subscribe();
     this.versionService.setReload(true);
   }
 
